feat(desktop): make explorer the marketplace view, drop design exploration harness
